Start implmenting in-memory node store
The Store is a glorified Map that persists Nodes. The goal with the
Store interface is to be implementable both as a memory-backed storage
and a file-backed storage.

Creating it led me to also create the
Node interface that gathers operations that all node types support.
The Node interface is missing a few methods right now, but they'll
be added as they are needed.