Improve hash text serialization
This changes the way that Qualified Hashes are serialized into text
so that the same format can be used for file names. I think it also
improves the readability of the resulting identifiers.

As part of this change, the logic that performs hashing was altered
to actually use the ContentLength of a HashDescriptor when choosing
the algorithm that it would use.