Add better tests for CacheStore
This involved creating a new Store operation, as the CacheStore needs
to ensure a level of consistency between the two layers of store that
it manages. The stores now have a CopyInto() method that causes one
store to replicate its content into another. The reason this approach
was taken over providing a way to iterate over the keys was to allow
stores to use their own internal representation of a given node ID as
a key. Providing a Keys() method would require all implementations to
use the same transformation from *fields.QualifiedHash to key type.