Implement using gpg2 to sign messages with encrypted keys
Rather than write code that handles PGP passphrases, this commit
adds support for invoking a system's installation of gpg2 to sign
forest nodes. This has the added advantage of leveraging a user's
existing GPG configuration, so they are prompted for their passphrase
with a familiar dialogue and their passphrase is cached according to
their gpg settings.

This commit also adds support for creating arbor forest nodes using
PGP keys with the `--gpguser` flag.