Implement and test tag-based serialization
The goal of this commit is to implement a more flexible and idiomatic approach
to serialization using reflection and struct tags instead of my homebrewed
interfaces and many levels of indirection. I think that the result is cleaner
and more maintainable.

The primary change is the implementation of the `ArborSerialize` method in
`serializer.go`. The algorithm that this implements is relatively simple:

Given a reflect.Value `v`:
1. Ensure `v` is a struct, dereference `v` if necessary
2. Find all fields of `v` that are tagged with the `arbor` tag (ignore all
   others)
3. Sort the tagged fields by their `order` attributes
4. Create an empty buffer of bytes, `b`
5. For each field `f` in the sorted fields:
  a. If `f` has the `recurse` attribute with a value of "always" or "serialize",
     recursively call `ArborSerialize` on the field and write the results into `b`
  b. Cast `f` to an `encoding.BinaryMarshaler` and marshal it into `b`
6. Return the contents of `b`

This change required a battery of supporting changes:

- `commonNode` is now `CommonNode`. This allows us to reflect on `CommonNode`s
embedded within `Identity` and other node types.
- The `Signature` field of CommonNode has been moved to a new `Trailer` type.
Now that the source of truth for serialization order is related to how the types
are structured, the `Signature` needs to be the last field present in the
high-order type (like `Identity`) in order to be the last field serialized.
- All types that are part of nodes have been augmented with `arbor` tags.

Additionally, this commit includes a test to demonstrate that the new algorithm
produces the same results as the old one. This did require a few tweaks to the
old algorithm due to the relocation of the `Signature` field. The test
serializes one of each type of node using both the new and old techniques. It
fails if the two binary results do not match for any node.

You can run that test with:

$ go test -run TestNewArborSerializer