Implement and test tag-based deserialization
This commit adds a tag-based deserialization implementation as the inverse of
ArborSerialize. It also adds a test to demonstrate that the result of
serializing and deserializing a node is identical to the original node.

This algorithm is very similar to ArborSerialize, but must handle dynamic-length
fields. The `ProgressiveBinaryUnmarshaler` interface allows deserializing
dynamic-length fields by providing a way to determine how many bytes were read
from the data after performing `UnmarshalBinary`.

The algorithm can be described as:

Given a reflect.Value `v` and a slice of bytes `b`:
1. Ensure `v` is a struct, dereference `v` if necessary
2. Find all fields of `v` that are tagged with the `arbor` tag (ignore all
   others)
3. Sort the tagged fields by their `order` attributes
5. For each field `f` in the sorted fields:
   a. If `f` has the `recurse` attribute with a value of "always" or "deserialize",
      recursively call `ArborDeserialize` with `(f,b)` and set `b` to the
      returned slice of bytes.
   b. Cast `f` to an `ProgressiveBinaryUnmarshaler` and unmarshal it from `b`
   c. Reslice `b` to remove the bytes consumed by deserializing `f`, which can
      be obtained with `f.BytesConsumed()`.
6. Return the unused contents of `b`

This commit introduces a test to ensure that, for any node `n`,
`ArborDeserialize(ArborSerialize(n)) == n`. This test can be run with:

$ go test -run TestNewArborSerializerSymmetry

Algebraically, TestNewArborSerializer demonstrates that the new serialization is
equivalent to the old one, and TestNewArborSerializerSymmetry demonstrates that
the new deserialization is the inverse of the new serialization. This is
sufficient to show that the new implementation of both operations is equivalent
to the previous implementation.