Remove old serialization logic from forest package
This commit strips out the previous serialization technique for nodes. As I hope
can be apparent from the diff, this is a dramatic simplification that allowed us
to discard a lot of spread-out logic.

In order to implement `VersionAndNodeTypeOf`, I had to introduce a new type that
was just the first two fields (to deserialize them independently from the rest
of a node). This caused a small cascade of changes as field names were updated,
but I think that the result is fairly clear for readers and it made
`VersionAndNodeTypeOf` much simpler.

As a consequence of removing the old serialization, the `TestNewArborSerializer`
test is no longer useful (it would now test the new implementation against
itself), so it has been removed.