Use new serialization for fields package
This commit removes the old, janky serialization that the fields package
provided and replaces it with the new approach. In order to be able to import
the new serialization code, I had to move it into its own subpackage. It is
satisfying that this is possible, because the serialization implementation is
completely independent of the types that it serializes. Yay loose coupling!

Again, I hope that the sheer volume of code that this allows us to get rid of is
compelling.