Find a more reliable terminal configuration for editor
So my first attempt used `st` to launch an editor, but this sometimes
(for reasons that I haven't figured out yet) exited uncleanly. `xterm`
exits reliably correctly, but it doesn't scale to my HiDPI screen. I
found the proper options for gnome-terminal (finally), so that's what
this is.

Note that I also had to configure my gpg pinentry program so that it
prompted me with an X11 window instead of a TTY because the viewer is
consuming keystrokes and that makes it impossible to reliably enter
passphrases.