Fix dynamic height accounting by forking CellView
This commit resolves a problem where the total message history would grow
but the total scrollable area would not. The problem seems to have been
that the views.ViewPort underlying the CellView wouldn't be properly
informed that the rendered text had gotten larger. To fix this, I had to
fork the CellView and add a step to its Draw() function. I hope to find
a better way to achieve this, but we shall see...