Add a String() method to QualifiedHash()
I'm fairly sure that there's no way for base64url-encoding the
node's ID to actually fail. The only reason this has an error
return value at all is because the text.Marshaler interface
requires one. This adds a way to get the string form of an ID
without going through the pain of handling the error that won't
even happen.