Implement specialized Get operations as wrappers
All of the more specialized Get() functions can be (inefficiently)
implemented as calls to the generic Get() function. For speed of
getting a working prototype, this is what I propose that we should
do for now. In the future, the Grove filesystem layout should allow
us to make getting nodes of a specific type a very efficient operation.

I've marked all of the functions with a bug tag that results from
not checking the type of the returned nodes. It was either that or
write specialized checking code that would just be scrapped later
anyway. This also saves the up-front cost of writing tests for these
functions right now, since they have no unique behavior. Obviously
they will need tests when their implementations diverge from the
generic operation though.