Fix fakeFS Create() behavior
The Create() method needs to behave as closely as possible
to the os.Create() function from the standard library.
Having it create a new file instead of truncating an existing
one made it difficult to test with error injection, since
it would throw away any existing errFile and create
a new fakeFile on every call to Create.

Since the actual grove implementation doesn't yet
need to use Truncate(), it was added to a private
interface in the test package instead of the grove.File
interface