Implement test for Add() serialization errors
Since the current implementation of MarshalBinary() for Nodes
can't fail unless someone breaks the struct tags on the nodes,
I had to implement forest.Node with a type that can actually
fail. Seemed like it was easy enough to be worth doing for that
last little line of test coverage.