feat: optimize redraw with rune slice
Previously the output of Render() was a string which needed
the nth() helper to find a particular rune. This was only
because I didn't know you could cast a string to a rune slice
directly. Using slice indexing is much faster, and results
in a measurable performance increase when moving the cursor.