feat: optimize cellview Draw() performance
This commit alters the CellView implementation that we copied from
tcell so that it only calls GetCell() on cells that are actually
visible from the viewport.

This makes a dramatic difference in runtime performance because
previously it would call GetCell() on every cell in the underlying
model. In our case, that was the entire chat history (not just what
was on screen at the time). As our chat history got longer, paying
this linear time increase got more and more expensive to render each
new 'frame' of text. By restricting the actual iteration of the
content here to just what is visible, we change the Draw() operation
to operate with a relatively small (thousands) constant number of
operations instead of millions that grow each time you get a message.

Profiling indicates that this change moves our performance bottleneck
from the Draw() phase (GetCell(), SetContent(), etc...) to the
Render() phase, which is where I always expected for it to be. I'm
hopeful that this change will alleviate most people's scrolling
performance issues.