bugfix: ensure bottom status bar updates with arrow keys
This required moving the key handling logic out of the vendored copy of
cellview into the HistoryWidget type. I think this is the right move,
as the CellView doesn't have the right information to do a lot of the
bookkeeping involved correctly. However, it makes us drift even further
from the upstream CellView type. I had harbored hopes of eventually
upstreaming our modifications, but this might kill those. Ah well.