feat: enable mouse scrolling to move cursor
I actually think that this might should pan the view instead,
but this is a start. That's why I introduced some pan* methods
on the HistoryWidget instead of just re-using the cursor* methods.
They are literally aliases right now, but I intend to change that.

As a side note, I think that this might need some form of mouse
acceleration in order to be practical, but I'm a little afraid of
trying to properly implement that.