ci/cd: configure goreleaser to build fewer artifacts
I tried to slim the output down to only those which I actually expect
to need in the near future. We don't have anyone on the more obscure
OSes, and building/storing those is not free. Happy to revisit this
if anyone needs a platform that we don't currently target.